create common function stream for fetchEventSource
